Enable tar diffing

+++ lib/dockly/deb.rb
@@ -6,8 +6,8 @@ class Dockly::Deb
 
   logger_prefix '[dockly deb]'
   dsl_attribute :package_name, :version, :release, :arch, :build_dir,
-                :pre_install, :post_install, :pre_uninstall, :post_uninstall,
-                :s3_bucket, :files, :app_user
+                :deb_build_dir, :pre_install, :post_install, :pre_uninstall,
+                :post_uninstall, :s3_bucket, :files, :app_user
 
   dsl_class_attribute :docker, Dockly::Docker
   dsl_class_attribute :foreman, Dockly::Foreman
@@ -15,7 +15,8 @@ class Dockly::Deb
   default_value :version, '0.0'
   default_value :release, '0'
   default_value :arch, 'x86_64'
-  default_value :build_dir, 'build/deb'
+  default_value :build_dir, 'build'
+  default_value :deb_build_dir, 'deb'
   default_value :files, []
   default_value :app_user, 'nobody'
 
@@ -24,8 +25,8 @@ class Dockly::Deb
   end
 
   def create_package!
-    ensure_present! :build_dir
-    FileUtils.mkdir_p(build_dir)
+    ensure_present! :build_dir, :deb_build_dir
+    FileUtils.mkdir_p(File.join(build_dir, deb_build_dir))
     FileUtils.rm(build_path) if File.exist?(build_path)
     debug "exporting #{package_name} to #{build_path}"
     build_package
@@ -46,8 +47,8 @@ class Dockly::Deb
   end
 
   def build_path
-    ensure_present! :build_dir
-    "#{build_dir}/#{output_filename}"
+    ensure_present! :build_dir, :deb_build_dir
+    File.join(build_dir, deb_build_dir, output_filename)
   end
 
   def exists?
@@ -63,7 +64,7 @@ class Dockly::Deb
 
   def upload_to_s3
     return if s3_bucket.nil?
-    create_package! unless File.exist?(build_path)
+    raise "Package wasn't created!" unless File.exist?(build_path)
     info "uploading package to s3"
     Dockly::AWS.s3.put_bucket(s3_bucket) rescue nil
     Dockly::AWS.s3.put_object(s3_bucket, s3_object_name, File.new(build_path))
@@ -81,16 +82,26 @@ class Dockly::Deb
     "#{package_name}_#{version}.#{release}_#{arch}.deb"
   end
 
+  def startup_script
+    scripts = []
+    bb = Dockly::BashBuilder.new
+    scripts << bb.normalize_for_dockly
+    scripts << bb.get_and_install_deb(s3_url, "/opt/dockly/#{File.basename(s3_url)}")
+
+    scripts.join("\n")
+  end
+
 private
   def build_package
     ensure_present! :package_name, :version, :release, :arch
 
     info "building #{package_name}"
     @dir_package = FPM::Package::Dir.new
-    add_docker(@dir_package)
     add_foreman(@dir_package)
     add_files(@dir_package)
     add_docker_auth_config(@dir_package)
+    add_docker(@dir_package)
+    add_startup_script(@dir_package)
 
     debug "converting to deb"
     @deb_package = @dir_package.convert(FPM::Package::Deb)
@@ -119,18 +130,6 @@ private
     package.attributes[:prefix] = nil
   end
 
-  def add_docker(package)
-    return if docker.nil?
-    info "adding docker image"
-    docker.generate!
-    return unless docker.registry.nil?
-    package.attributes[:prefix] = docker.package_dir
-    Dir.chdir(File.dirname(docker.tar_path)) do
-      package.input(File.basename(docker.tar_path))
-    end
-    package.attributes[:prefix] = nil
-  end
-
   def add_files(package)
     return if files.empty?
     info "adding files to package"
@@ -144,13 +143,70 @@ private
   end
 
   def add_docker_auth_config(package)
-    return if docker.nil? || (registry = docker.registry).nil? || !registry.authentication_required?
+    return if (registry = get_registry).nil? || !registry.authentication_required?
     info "adding docker config file"
     registry.generate_config_file!
 
-    package.attributes[:prefix] = docker.auth_config_file || "~#{app_user}/.dockercfg"
+    package.attributes[:prefix] = registry.auth_config_file || "~#{app_user}"
     Dir.chdir(File.dirname(registry.config_file)) do
       package.input(File.basename(registry.config_file))
+    end
+    package.attributes[:prefix] = nil
+  end
+
+  def add_docker(package)
+    return if docker.nil? || docker.s3_bucket
+    info "adding docker image"
+    docker.generate!
+    return unless docker.registry.nil?
+    package.attributes[:prefix] = docker.package_dir
+    Dir.chdir(File.dirname(docker.tar_path)) do
+      package.input(File.basename(docker.tar_path))
+    end
+    package.attributes[:prefix] = nil
+  end
+
+  def get_registry
+    if docker && registry = docker.registry
+      registry
+    end
+  end
+
+  def post_startup_script
+    scripts = ["#!/bin/bash"]
+    bb = Dockly::BashBuilder.new
+    scripts << bb.normalize_for_dockly
+    if get_registry
+      scripts << bb.registry_import(docker.repo, docker.tag)
+    elsif docker
+      if docker.s3_bucket.nil?
+        docker_output = File.join(docker.package_dir, docker.export_filename)
+        if docker.tar_diff
+          scripts << bb.file_diff_docker_import(docker.import, docker_output, docker.name, docker.tag)
+        else
+          scripts << bb.file_docker_import(docker_output, docker.name, docker.tag)
+        end
+      else
+        if docker.tar_diff
+          scripts << bb.s3_diff_docker_import(docker.import, docker.s3_url, docker.name, docker.tag)
+        else
+          scripts << bb.s3_docker_import(docker.s3_url, docker.name, docker.tag)
+        end
+      end
+    end
+    scripts.join("\n")
+  end
+
+  def add_startup_script(package)
+    ensure_present! :build_dir
+    startup_script_path = File.join(build_dir, "dockly-startup.sh")
+    File.open(startup_script_path, 'w+') do |f|
+      f.write(post_startup_script)
+      f.chmod(0755)
+    end
+    package.attributes[:prefix] = "/opt/dockly"
+    Dir.chdir(build_dir) do
+      package.input("dockly-startup.sh")
     end
     package.attributes[:prefix] = nil
   end
